Rename some components and containing folders
This is done in order for better conventions in naming folders
and page components.

The top level folder name now refers to the action rather than
the page itself. Eg., the folder containing the AddEntryPage
component is called addEntry instead of addEntryPage. Similarly
the Leaderboard component is now called LeaderBoardPage.