Remove the JSScheme config from Info.plist
Having two places to store config - one to be used by native code and
another to be used by the JS code was cumbersome. A JSON config file
makes more sense to store global config since it is more compact and
has wider acceptance. It can be easily used in the JS code as well.

- Use appConfig.json to store the JS scheme
- Add appConfig.json as a resource to the iOS app bundle
- In AppDelegate.m use this JSON config to determine the JSScheme