iOS | Remove the staging and dev targets
Targets should not be used for configuration. Earlier we
had used different targets for each configuration to have
different bundle identifiers. This is not a good pratice as
certain settings like the team have to be managed for each
target.

To have different bundle IDs, we now have different values for
the build setting PRODUCT_BUNDLE_IDENTIFIER for different
configurations. This is then used to set the bundle id.